major directory refactoring
