--- Vagrantfile
@@ -1,66 +1,0 @@
-# -*- mode: ruby -*-
-# vi: set ft=ruby :
-
-$ip_file = "db_ip.txt"
-
-Vagrant.configure("2") do |config|
-  config.vm.box = 'digital_ocean'
-  config.vm.box_url = "https://github.com/devopsgroup-io/vagrant-digitalocean/raw/master/box/digital_ocean.box"
-  config.ssh.private_key_path = '~/.ssh/id_rsa'
-  config.vm.synced_folder ".", "/vagrant", type: "rsync", rsync__exclude: ["./node_modules*","./package-lock.json","./out.log"]
-
-  config.vm.define "webserver", primary: true do |server|
-    server.vm.provider :digital_ocean do |provider|
-      provider.ssh_key_name = ENV["SSH_KEY_NAME"]
-      provider.token = ENV["DIGITAL_OCEAN_TOKEN"]
-      provider.image = 'ubuntu-22-04-x64'
-      provider.region = 'fra1'
-      provider.size = 's-1vcpu-1gb'
-      provider.privatenetworking = true
-    end
-
-    server.vm.hostname = "webserver"
-
-    server.trigger.after :up do |trigger|
-      trigger.info =  "Writing webserber's IP to file..."
-      trigger.ruby do |env,machine|
-        remote_ip = machine.instance_variable_get(:@communicator).instance_variable_get(:@connection_ssh_info)[:host]
-        File.write($ip_file, remote_ip)
-      end
-    end
-
-    server.vm.provision "shell", inline: <<-SHELL
-      export DB_IP=`cat /vagrant/db_ip.txt`
-      echo $DB_IP
-
-      echo "***********************"
-      echo "Updating apt sources..."
-      echo "***********************"
-      
-      
-      wget https://raw.githubusercontent.com/nvm-sh/nvm/v0.39.5/install.sh
-      chmod +x ./install.sh
-      bash ./install.sh -b
-      
-      export NVM_DIR="$HOME/.nvm"
-      [ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"
-      [ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion"
-
-      source ~/.bashrc 
-      source ~/.nvm/nvm.sh
-
-      nvm install 14 
-      cp -r /vagrant/* $HOME     
-      npm install
-      npm start > out.log &
-      echo "================================================================="
-      echo "=                            DONE                               ="
-      echo "================================================================="
-      echo "Navigate in your browser to:"
-      THIS_IP=`hostname -I | cut -d" " -f1`
-      echo "http://${THIS_IP}:5000"
-    SHELL
-  end
-
-    
-end
