+++ app/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "minitwit-2",
+  "version": "0.0.0",
+  "private": true,
+  "scripts": {
+    "start": "node ./bin/www",
+    "start-api": "node ./API/api.js",
+    "test": "mocha ./tests/*.test.js"
+  },
+  "dependencies": {
+    "bcrypt": "^5.1.1",
+    "cookie-parser": "~1.4.4",
+    "crypto-js": "^4.2.0",
+    "debug": "~2.6.9",
+    "ejs": "~2.6.1",
+    "express": "~4.16.1",
+    "express-ejs-layouts": "^2.5.1",
+    "express-flash": "^0.0.2",
+    "express-session": "^1.18.0",
+    "http-errors": "~1.6.3",
+    "morgan": "~1.9.1",
+    "os-utils": "^0.0.14",
+    "pg": "^8.11.3",
+    "prom-client": "^15.1.0",
+    "sqlite3": "^5.1.7",
+    "uuid": "^9.0.1"
+  },
+  "devDependencies": {
+    "assert": "^2.1.0",
+    "chai": "^5.1.0",
+    "mocha": "~3.5.3",
+    "supertest": "^6.3.4"
+  }
+}
