+++ app/src/views/register.ejs
