--- routes/EXPRESS_ROUTE_index.js
@@ -1,12 +1,0 @@
-// This is how you do routes. But we are not using that at the moment. [LAKJ]
-
-
-// var express = require('express');
-// var router = express.Router();
-
-// /* GET home page. */
-// router.get('/', function(req, res, next) {
-//   res.render('index', { title: 'Express' });
-// });
-
-// module.exports = router;
